refactor: extract the render tree into src/ui/
Every render fn already took &App, so the whole tree lifts out of main.rs
with a one-way dependency: ui reads App, writes FrameOut, mutates nothing.

One module per screen — a contributor changing the queue opens ui/queue.rs.
ui stays in the binary crate (mod ui, not lib.rs) because it needs App, which
stays in main.rs for now.

Pure code motion: no logic edits, visibility widened to pub(crate) only.
