+++ src/main.rs
@@ -5,6 +5,10 @@
 //! Multi-section library (playlists / liked / albums / artists), shuffle, repeat,
 //! and a live queue view.
 
+/// The render tree. Reads `App`, writes `FrameOut`; never the other way round.
+/// Lives in the binary (not the library) because it needs `App`, which is here.
+mod ui;
+
 use std::io;
 use std::sync::{Arc, Mutex};
 use std::time::{Duration, Instant};
@@ -36,6 +40,8 @@ use myx::term::{acquire_single_instance_lock, init_terminal, restore_terminal, T
 use myx::theme::{Theme, TOKYONIGHT};
 use myx::util::{center_v, fmt_ms, track_id_from_uri, truncate, uri_to_url, urlencode, vol_u16};
 use myx::webapi::WebApi;
+use ui::{render, render_loading};
+
 use souvlaki::{
     MediaControlEvent, MediaControls, MediaMetadata, MediaPlayback, MediaPosition, PlatformConfig,
     SeekDirection,
@@ -3554,956 +3560,6 @@ fn spawn_restore(
 
 // --- Live playback state (server-side) end ---
 
-// ------------------------------------------------------------------ render
-
-fn render(f: &mut Frame, app: &App, out: &mut FrameOut, repaint: ArtRepaint) {
-    let theme = app.theme.displayed;
-    let area = f.area();
-    f.render_widget(Block::default().style(theme.base()), area);
-    let area = area.inner(Margin::new(2, 1));
-
-    let rows = Layout::vertical([
-        Constraint::Length(1), // header
-        Constraint::Length(1), // spacer
-        Constraint::Min(6),    // body (library | active view)
-        Constraint::Length(1), // spacer
-        Constraint::Length(2), // now-playing strip
-        Constraint::Length(1), // footer
-    ])
-    .split(area);
-
-    // Header: wordmark + view tabs (right-aligned) + status.
-    // Fullwidth wordmark (each letter = 2 cells) reads as a bigger "myx"
-    // than the terminal font allows on a single row; bolded for weight.
-    let mut header: Vec<Span> =
-        gradient_line("\u{FF2D}\u{FF39}\u{FF38}", &[theme.primary, theme.accent])
-            .into_iter()
-            .map(|mut sp| {
-                sp.style = sp.style.add_modifier(Modifier::BOLD);
-                sp
-            })
-            .collect();
-    if !app.status.is_empty() {
-        header.push(Span::styled(format!("   {}", app.status), theme.muted()));
-    }
-    f.render_widget(Paragraph::new(Line::from(header)), rows[0]);
-    f.render_widget(
-        Paragraph::new(Line::from(view_tabs(app, theme))).alignment(Alignment::Right),
-        rows[0],
-    );
-    // Per-tab hit rects for the mouse (mirrors view_tabs: "\u2190\u2192 " prefix + labels joined by " \u00b7 ").
-    let mut total: usize = 3; // "\u2190\u2192 "
-    for (i, v) in RightView::ALL.iter().enumerate() {
-        if i > 0 {
-            total += 3;
-        } // " \u00b7 "
-        total += v.label().chars().count();
-    }
-    let mut tx_x = rows[0]
-        .right()
-        .saturating_sub(total as u16)
-        .saturating_add(3);
-    let mut tabs = Vec::with_capacity(RightView::ALL.len());
-    for (i, v) in RightView::ALL.iter().enumerate() {
-        if i > 0 {
-            tx_x = tx_x.saturating_add(3);
-        }
-        let w = v.label().chars().count() as u16;
-        tabs.push((
-            *v,
-            Rect {
-                x: tx_x,
-                y: rows[0].y,
-                width: w,
-                height: 1,
-            },
-        ));
-        tx_x = tx_x.saturating_add(w);
-    }
-    out.hits.tabs = tabs;
-
-    let right = if app.view.zen {
-        // Hidden, not zero-width: a rendered sidebar still claims mouse rects.
-        out.hits.lib = None;
-        out.hits.scroll = None;
-        rows[2]
-    } else {
-        let body = Layout::horizontal([Constraint::Percentage(30), Constraint::Min(24)])
-            .spacing(3)
-            .split(rows[2]);
-        render_library(f, app, out, theme, body[0]);
-        body[1]
-    };
-    match app.view.mode {
-        RightView::NowPlaying => render_nowplaying_view(f, app, theme, right, repaint),
-        RightView::Lyrics => render_lyrics(f, app, theme, right),
-        RightView::Queue => render_queue_view(f, app, theme, right),
-    }
-
-    render_now_strip(f, app, out, theme, rows[4]);
-    render_footer(f, app, theme, rows[5]);
-
-    if app.view.actions.is_some() {
-        render_actions_overlay(f, app, theme, area);
-    }
-}
-
-/// Blank `area` and force it out to the terminal, erasing an inline image.
-///
-/// `ratatui-image` marks the image's cells `Skip` without changing their
-/// symbols, and a blank cell compares equal to the blank that was already there
-/// — so an ordinary `Clear` writes nothing and the picture survives it.
-/// `AlwaysUpdate` is what makes the diff emit them anyway.
-fn wipe_area(f: &mut Frame, area: Rect) {
-    let buf = f.buffer_mut();
-    for y in area.top()..area.bottom() {
-        for x in area.left()..area.right() {
-            if let Some(cell) = buf.cell_mut((x, y)) {
-                cell.reset();
-                cell.set_diff_option(CellDiffOption::AlwaysUpdate);
-            }
-        }
-    }
-}
-
-/// Keep the terminal's own pixels in `area` by telling the diff to leave those
-/// cells alone — what `ratatui-image` does for the image it just drew, without
-/// drawing it again.
-fn hold_area(f: &mut Frame, area: Rect) {
-    let buf = f.buffer_mut();
-    for y in area.top()..area.bottom() {
-        for x in area.left()..area.right() {
-            if let Some(cell) = buf.cell_mut((x, y)) {
-                cell.set_diff_option(CellDiffOption::Skip);
-            }
-        }
-    }
-}
-
-/// Force whatever is already in `area` out to the terminal, so an overlay lands
-/// on top of an inline image instead of being skipped over it.
-fn force_area(f: &mut Frame, area: Rect) {
-    let buf = f.buffer_mut();
-    for y in area.top()..area.bottom() {
-        for x in area.left()..area.right() {
-            if let Some(cell) = buf.cell_mut((x, y)) {
-                cell.set_diff_option(CellDiffOption::AlwaysUpdate);
-            }
-        }
-    }
-}
-
-/// The `Now Playing · Lyrics · Visualizer` indicator, active one lit.
-fn view_tabs<'a>(app: &App, theme: Theme) -> Vec<Span<'a>> {
-    let mut spans = vec![Span::styled("←→ ", theme.muted())];
-    for (i, v) in RightView::ALL.iter().enumerate() {
-        if i > 0 {
-            spans.push(Span::styled(" · ", theme.muted()));
-        }
-        let style = if *v == app.view.mode {
-            Style::default()
-                .fg(theme.primary.into())
-                .add_modifier(Modifier::BOLD)
-        } else {
-            theme.muted()
-        };
-        spans.push(Span::styled(v.label(), style));
-    }
-    spans
-}
-
-/// Where the list viewport starts, given where it started last frame.
-///
-/// Keeps `margin` rows visible either side of the cursor (vim's `scrolloff`).
-/// Threading the previous `offset` back in is what makes it sticky: the cursor
-/// moves freely inside the window, which follows only when pushed.
-fn scroll_offset(offset: usize, selected: usize, cap: usize, total: usize, margin: usize) -> usize {
-    if cap == 0 || total <= cap {
-        return 0;
-    }
-    // Wider than half the viewport and the two bounds below would fight.
-    let margin = margin.min((cap - 1) / 2);
-    offset
-        .min(selected.saturating_sub(margin)) // cursor near the top → scroll up
-        .max((selected + margin + 1).saturating_sub(cap)) // near the bottom → down
-        .min(total - cap)
-}
-
-fn render_library(f: &mut Frame, app: &App, out: &mut FrameOut, theme: Theme, area: Rect) {
-    f.render_widget(Block::default().style(theme.panel()), area);
-    let inner = area.inner(Margin::new(2, 1));
-    if inner.height < 2 {
-        return;
-    }
-
-    // Header line: drill-in title, search input/results, or section indicator.
-    let head: Line = if let Some(d) = app.browse.details.last() {
-        Line::from(vec![
-            Span::styled("‹ ", Style::default().fg(theme.primary.into())),
-            Span::styled(
-                truncate(&d.title, inner.width.saturating_sub(8) as usize),
-                theme.heading(),
-            ),
-            Span::styled("  Esc", theme.muted()),
-        ])
-    } else if app.search.input_mode {
-        Line::from(vec![
-            Span::styled("search: ", theme.heading()),
-            Span::styled(
-                format!("{}▏", app.search.query),
-                Style::default().fg(theme.text.into()),
-            ),
-        ])
-    } else if app.search.searching {
-        Line::from(vec![
-            Span::styled("search: ", theme.heading()),
-            Span::styled(
-                app.search.query.clone(),
-                Style::default().fg(theme.text.into()),
-            ),
-            Span::styled("  (Esc)", theme.muted()),
-        ])
-    } else {
-        let mut spans = vec![
-            Span::styled("‹ ", theme.muted()),
-            Span::styled(app.browse.section.label(), theme.heading()),
-            Span::styled(" ›", theme.muted()),
-            Span::styled(
-                format!(
-                    "  {}/{} · {}",
-                    app.browse.section.index() + 1,
-                    Section::ALL.len(),
-                    app.cur_items().len()
-                ),
-                theme.muted(),
-            ),
-        ];
-        if app.browse.sort != SortMode::Added {
-            spans.push(Span::styled(
-                format!("  ⇅{}", app.browse.sort.label()),
-                Style::default().fg(theme.accent.into()),
-            ));
-        }
-        Line::from(spans)
-    };
-    f.render_widget(
-        Paragraph::new(head).block(Block::default().style(theme.panel())),
-        Rect {
-            x: inner.x,
-            y: inner.y,
-            width: inner.width,
-            height: 1,
-        },
-    );
-
-    let list_top = inner.y + 2;
-    if list_top >= inner.bottom() {
-        return;
-    }
-    let cap = (inner.bottom() - list_top) as usize;
-    let total_items = app.cur_items().len();
-
-    if total_items == 0 {
-        out.hits.scroll = None;
-        out.hits.lib = None;
-        f.render_widget(
-            Paragraph::new(Line::from(Span::styled("(empty)", theme.muted())))
-                .block(Block::default().style(theme.panel())),
-            Rect {
-                x: inner.x,
-                y: list_top,
-                width: inner.width,
-                height: 1,
-            },
-        );
-        return;
-    }
-
-    let offset = scroll_offset(
-        out.lib_offset,
-        app.browse.selected,
-        cap,
-        total_items,
-        myx::config::get().scrolloff,
-    );
-    out.hits.lib = Some(Rect {
-        x: inner.x,
-        y: list_top,
-        width: inner.width,
-        height: cap as u16,
-    });
-    out.lib_offset = offset;
-    let overflow = total_items > cap && inner.width > 2;
-    // Reserve an extra gutter column for the scrollbar (+1 char of padding).
-    let max = inner.width.saturating_sub(if overflow { 3 } else { 2 }) as usize;
-
-    let items = app.cur_items();
-    for (row, item) in items.iter().skip(offset).take(cap).enumerate() {
-        let idx = offset + row;
-        let y = list_top + row as u16;
-        let rect = Rect {
-            x: inner.x,
-            y,
-            width: inner.width,
-            height: 1,
-        };
-
-        // Header rows: a bold section label (Home feed groups), not selectable.
-        if item.is_header {
-            f.render_widget(
-                Paragraph::new(Line::from(Span::styled(
-                    item.name.clone(),
-                    Style::default()
-                        .fg(theme.accent.into())
-                        .add_modifier(Modifier::BOLD),
-                )))
-                .block(Block::default().style(theme.panel())),
-                rect,
-            );
-            continue;
-        }
-
-        let selected = idx == app.browse.selected;
-        let bg = if selected {
-            theme.background_element.into()
-        } else {
-            theme.background_panel.into()
-        };
-        let block = left_bar_block(&theme, selected, bg);
-        let style = if selected {
-            Style::default()
-                .fg(theme.text.into())
-                .add_modifier(Modifier::BOLD)
-        } else {
-            theme.muted()
-        };
-        // Mark rows that `P` can play outright (playlist / album / artist), so
-        // they're distinguishable from tracks at a glance.
-        let playable_ctx = context_target(item).is_some() && !item.is_play;
-        let max = if playable_ctx {
-            max.saturating_sub(2)
-        } else {
-            max
-        };
-        let label = truncate(&item.name, max);
-        let mut spans = Vec::new();
-        if playable_ctx {
-            spans.push(Span::styled(
-                " ▶",
-                Style::default().fg(theme.border_dimmest.into()),
-            ));
-        }
-        spans.push(Span::styled(format!(" {label}"), style));
-        if !item.subtitle.is_empty() {
-            let used = label.chars().count() + 1;
-            let room = max.saturating_sub(used + 3);
-            if room > 3 {
-                spans.push(Span::styled(
-                    " · ",
-                    Style::default().fg(theme.border_dimmest.into()),
-                ));
-                spans.push(Span::styled(
-                    truncate(&item.subtitle, room),
-                    theme.muted().add_modifier(Modifier::DIM),
-                ));
-            }
-        }
-        f.render_widget(Paragraph::new(Line::from(spans)).block(block), rect);
-    }
-
-    // Subtle scrollbar: a hairline 1/8 track with a slightly denser 1/4 thumb,
-    // in the right gutter. Shown only on overflow; the track rect is stashed on
-    // `app` so mouse drags can scroll it.
-    if overflow {
-        let total = total_items;
-        let sb_x = inner.right();
-        let track_h = cap;
-        let thumb_h = (cap * cap).div_ceil(total).max(1).min(track_h);
-        let travel = track_h - thumb_h;
-        let max_off = total - cap;
-        let thumb_y0 = (offset * travel + max_off / 2)
-            .checked_div(max_off)
-            .unwrap_or(0);
-        for i in 0..track_h {
-            let y = list_top + i as u16;
-            if y >= inner.bottom() {
-                break;
-            }
-            let in_thumb = i >= thumb_y0 && i < thumb_y0 + thumb_h;
-            let (glyph, color) = if in_thumb {
-                ("\u{258E}", theme.text_muted) // 1/4 block - thumb
-            } else {
-                ("\u{258F}", theme.border_dimmest) // 1/8 block - track
-            };
-            f.render_widget(
-                Paragraph::new(Span::styled(glyph, Style::default().fg(color.into()))),
-                Rect {
-                    x: sb_x,
-                    y,
-                    width: 1,
-                    height: 1,
-                },
-            );
-        }
-        out.hits.scroll = Some(Rect {
-            x: sb_x,
-            y: list_top,
-            width: 1,
-            height: track_h as u16,
-        });
-        out.hits.scroll_len = total;
-    } else {
-        out.hits.scroll = None;
-    }
-}
-
-/// View ①: album art with track details directly beneath — centered as a group.
-fn render_nowplaying_view(f: &mut Frame, app: &App, theme: Theme, area: Rect, repaint: ArtRepaint) {
-    if app.playback.now.is_none() {
-        f.render_widget(
-            Paragraph::new("Nothing playing.\nBrowse ← and press Enter.")
-                .style(theme.muted())
-                .alignment(Alignment::Center),
-            center_v(area, 2),
-        );
-        return;
-    }
-
-    // Split: album art + track info on top, a compact spectrum below, lifted a
-    // little off the bottom.
-    let chunks = Layout::vertical([
-        Constraint::Min(6),    // art + text
-        Constraint::Length(7), // spectrum
-        Constraint::Length(2), // breathing room (lifts the spectrum up)
-    ])
-    .split(area);
-    let top = chunks[0];
-    // Push the art + info group down a little from the top.
-    let top = Rect {
-        x: top.x,
-        y: top.y + 3,
-        width: top.width,
-        height: top.height.saturating_sub(3),
-    };
-    let viz_area = chunks[1];
-
-    // Derive the cover's cell footprint from the terminal's font aspect so a
-    // square image renders square (and our centering math is exact).
-    let font = app.svc.picker.font_size();
-    let fw = font.width.max(1) as u32;
-    let fh = font.height.max(1) as u32;
-
-    // Reserve 3 rows for text (+1 gap). Cap the art so the group stays compact.
-    let avail_h = top.height.saturating_sub(4);
-    let mut art_h = avail_h.clamp(3, 14);
-    // Square image width in cells for this height: w = h * fh / fw.
-    let mut art_w = (art_h as u32 * fh / fw) as u16;
-    if art_w > top.width {
-        art_w = top.width;
-        art_h = (art_w as u32 * fw / fh) as u16;
-    }
-
-    let group_h = art_h + 4; // art + gap + title + artist + album
-    let art_y = top.y + top.height.saturating_sub(group_h) / 2;
-    let art_x = top.x + top.width.saturating_sub(art_w) / 2;
-    let art_rect = Rect {
-        x: art_x,
-        y: art_y,
-        width: art_w,
-        height: art_h,
-    };
-
-    match app.playback.now.as_ref().and_then(|n| n.cover.as_ref()) {
-        _ if repaint == ArtRepaint::Wipe => wipe_area(f, art_rect),
-        // Writing the escape means transmitting the image, so only do it when
-        // something actually asked for it. A theme fade repaints every glyph on
-        // screen dozens of times, and re-sending the cover on each of those is
-        // what made it flicker.
-        Some(cover) if repaint == ArtRepaint::Draw || cover.needs_send(art_rect) => {
-            cover.render(f, art_rect)
-        }
-        // Already on screen: hold the cells so nothing overwrites the picture,
-        // and send nothing.
-        Some(_) => hold_area(f, art_rect),
-        None => wipe_area(f, art_rect),
-    }
-
-    if let Some(n) = app.playback.now.as_ref() {
-        let text_rect = Rect {
-            x: top.x,
-            y: art_rect.y + art_h + 1,
-            width: top.width,
-            height: 3,
-        };
-        let lines = vec![
-            Line::from(Span::styled(
-                truncate(&n.title, top.width as usize),
-                Style::default()
-                    .fg(theme.text.into())
-                    .add_modifier(Modifier::BOLD),
-            )),
-            Line::from(Span::styled(
-                truncate(&n.artist, top.width as usize),
-                Style::default().fg(theme.primary.into()),
-            )),
-            Line::from(Span::styled(
-                truncate(&n.album, top.width as usize),
-                theme.muted(),
-            )),
-        ];
-        f.render_widget(
-            Paragraph::new(lines).alignment(Alignment::Center),
-            text_rect,
-        );
-    }
-
-    render_visualizer(f, app, theme, viz_area);
-}
-
-/// Slim persistent bottom strip: play state + track, then the progress bar.
-fn render_now_strip(f: &mut Frame, app: &App, out: &mut FrameOut, theme: Theme, area: Rect) {
-    let rows = Layout::vertical([Constraint::Length(1), Constraint::Length(1)]).split(area);
-
-    // Volume meter (top row, far right). Still honest in remote mode: +/- goes
-    // to the remote device's volume.
-    render_volume(f, app, out, theme, rows[0]);
-
-    // Seek/progress bar (bottom row). Record bar geometry for click-to-seek.
-    let pos = app.playback.position_ms();
-    let left_len = format!("{} ", fmt_ms(pos)).chars().count() as u16;
-    let right_len = format!(
-        " {}",
-        fmt_ms(
-            app.playback
-                .now
-                .as_ref()
-                .map(|n| n.duration_ms)
-                .unwrap_or(0)
-        )
-    )
-    .chars()
-    .count() as u16;
-    let bar_w = rows[1].width.saturating_sub(left_len + right_len);
-    out.hits.bar = Some(Rect {
-        x: rows[1].x + left_len,
-        y: rows[1].y,
-        width: bar_w,
-        height: 1,
-    });
-    render_progress(f, app, theme, rows[1]);
-}
-
-/// The volume meter — a graduated ramp + percentage, right-aligned in `area`.
-/// Stashes the 8-bar region on `out` for click/drag control.
-fn render_volume(f: &mut Frame, app: &App, out: &mut FrameOut, theme: Theme, area: Rect) {
-    const VLEV: [char; 8] = ['▁', '▂', '▃', '▄', '▅', '▆', '▇', '█'];
-    let filled = (app.transport.volume as usize * VLEV.len() + 50) / 100;
-    let mut vspans: Vec<Span> = Vec::with_capacity(VLEV.len() + 1);
-    for (i, ch) in VLEV.iter().enumerate() {
-        let color = if i < filled {
-            theme.primary
-        } else {
-            theme.border_dimmest
-        };
-        vspans.push(Span::styled(
-            ch.to_string(),
-            Style::default().fg(color.into()),
-        ));
-    }
-    vspans.push(Span::styled(
-        format!(" {:>3}%", app.transport.volume),
-        theme.muted(),
-    ));
-    f.render_widget(
-        Paragraph::new(Line::from(vspans)).alignment(Alignment::Right),
-        area,
-    );
-    // 8-bar region for click/drag. Content is 13 cells (8 bars + " NNN%"),
-    // right-aligned, so the bars start 13 cells in from the right edge.
-    out.hits.vol = Some(Rect {
-        x: area.right().saturating_sub(13),
-        y: area.y,
-        width: VLEV.len() as u16,
-        height: 1,
-    });
-}
-
-fn render_lyrics(f: &mut Frame, app: &App, theme: Theme, area: Rect) {
-    let inner = area.inner(Margin::new(2, 0));
-    if inner.height == 0 {
-        return;
-    }
-    let max = inner.width as usize;
-
-    // Header: current track title + "artist · album", above the lyrics.
-    let mut lyrics_area = inner;
-    if let Some(n) = app.playback.now.as_ref() {
-        let head = Layout::vertical([
-            Constraint::Length(1), // title
-            Constraint::Length(1), // artist / album
-            Constraint::Length(1), // spacer
-            Constraint::Min(1),    // lyrics
-        ])
-        .split(inner);
-        f.render_widget(
-            Paragraph::new(Line::from(Span::styled(
-                truncate(&n.title, max),
-                Style::default()
-                    .fg(theme.text.into())
-                    .add_modifier(Modifier::BOLD),
-            )))
-            .alignment(Alignment::Center),
-            head[0],
-        );
-        let sub = if n.album.is_empty() {
-            n.artist.clone()
-        } else {
-            format!("{} · {}", n.artist, n.album)
-        };
-        f.render_widget(
-            Paragraph::new(Line::from(Span::styled(truncate(&sub, max), theme.muted())))
-                .alignment(Alignment::Center),
-            head[1],
-        );
-        lyrics_area = head[3];
-    }
-
-    if app.view.lyrics.is_empty() {
-        let msg = if app.playback.now.is_some() {
-            "♪︎  no lyrics for this track"
-        } else {
-            "♪︎  nothing playing"
-        };
-        f.render_widget(
-            Paragraph::new(msg)
-                .style(theme.muted())
-                .alignment(Alignment::Center),
-            center_v(lyrics_area, 1),
-        );
-        return;
-    }
-
-    let h = lyrics_area.height as usize;
-    let pos = app.playback.position_ms();
-    let cur = if app.view.lyrics_synced {
-        app.view
-            .lyrics
-            .iter()
-            .rposition(|(t, _)| *t <= pos)
-            .unwrap_or(0)
-    } else {
-        0
-    };
-    let start = cur.saturating_sub(h / 2);
-
-    let mut lines: Vec<Line> = Vec::with_capacity(h);
-    for (i, (_, text)) in app.view.lyrics.iter().enumerate().skip(start).take(h) {
-        let style = if app.view.lyrics_synced && i == cur {
-            Style::default()
-                .fg(theme.primary.into())
-                .add_modifier(Modifier::BOLD)
-        } else if app.view.lyrics_synced && i < cur {
-            Style::default().fg(theme.border_subtle.into())
-        } else {
-            theme.muted()
-        };
-        let txt = if text.is_empty() {
-            "♪︎".to_string()
-        } else {
-            truncate(text, max)
-        };
-        lines.push(Line::from(Span::styled(txt, style)));
-    }
-    f.render_widget(
-        Paragraph::new(lines).alignment(Alignment::Center),
-        lyrics_area,
-    );
-}
-
-fn render_visualizer(f: &mut Frame, app: &App, theme: Theme, area: Rect) {
-    let active = app
-        .svc
-        .engine
-        .bands
-        .try_lock()
-        .map(|g| g.is_active)
-        .unwrap_or(false);
-    if !active {
-        return;
-    }
-    let Ok(guard) = app.svc.engine.bands.try_lock() else {
-        return;
-    };
-    let values: [f32; NUM_BANDS] = guard.values;
-    let peak = guard.peak_envelope.max(1e-6);
-    drop(guard);
-
-    // Cap the spectrum to a centered band — full-pane bars are too tall/wide.
-    let vh = ((area.height as u32 * 3 / 5) as u16)
-        .clamp(6, 14)
-        .min(area.height);
-    let vw = ((area.width as u32 * 9 / 10) as u16)
-        .clamp(24, 80)
-        .min(area.width);
-    let vrect = Rect {
-        x: area.x + area.width.saturating_sub(vw) / 2,
-        y: area.y + area.height.saturating_sub(vh) / 2,
-        width: vw,
-        height: vh,
-    };
-    let w = vrect.width as usize;
-    let h = vrect.height as usize;
-    if w == 0 || h == 0 {
-        return;
-    }
-
-    // 1. Box-average the bands into each column (anti-aliasing vs. single-pick).
-    let mut cols = vec![0.0f32; w];
-    for (x, c) in cols.iter_mut().enumerate() {
-        let lo = x * NUM_BANDS / w;
-        let hi = (((x + 1) * NUM_BANDS / w).max(lo + 1)).min(NUM_BANDS);
-        let sum: f32 = values[lo..hi].iter().sum();
-        let v = sum / (hi - lo) as f32;
-        // Perceptual curve so quiet detail stays visible.
-        *c = (v / peak).sqrt().clamp(0.0, 1.0);
-    }
-
-    // 2. Spatial smoothing — a couple of weighted passes so the envelope flows
-    //    instead of spiking. This is what kills the "chopped" look.
-    for _ in 0..2 {
-        let src = cols.clone();
-        for x in 0..w {
-            let l = src[x.saturating_sub(1)];
-            let r = src[(x + 1).min(w - 1)];
-            cols[x] = l * 0.25 + src[x] * 0.5 + r * 0.25;
-        }
-    }
-
-    // 3. Render with an eighth-block sub-cell tip and a vertical color gradient
-    //    (info at the base → primary → accent at the peaks) for a smooth wash.
-    const LEVELS: [char; 8] = ['▁', '▂', '▃', '▄', '▅', '▆', '▇', '█'];
-    let stops = [theme.info, theme.primary, theme.accent];
-
-    let mut lines: Vec<Line> = Vec::with_capacity(h);
-    for row in 0..h {
-        let from_bottom = (h - 1 - row) as f32;
-        let vfrac = if h > 1 {
-            from_bottom / (h - 1) as f32
-        } else {
-            0.0
-        };
-        let color: ratatui::style::Color = gradient::interpolate(&stops, vfrac).into();
-        let mut spans: Vec<Span> = Vec::with_capacity(w);
-        for &v in &cols {
-            let filled = v * h as f32 - from_bottom;
-            let ch = if filled >= 1.0 {
-                '█'
-            } else if filled <= 0.0 {
-                ' '
-            } else {
-                LEVELS[((filled * 8.0) as usize).clamp(1, 8) - 1]
-            };
-            if ch == ' ' {
-                spans.push(Span::raw(" "));
-            } else {
-                spans.push(Span::styled(ch.to_string(), Style::default().fg(color)));
-            }
-        }
-        lines.push(Line::from(spans));
-    }
-    f.render_widget(Paragraph::new(lines), vrect);
-}
-
-fn render_progress(f: &mut Frame, app: &App, theme: Theme, area: Rect) {
-    let (pos, dur) = match &app.playback.now {
-        Some(n) => (app.playback.position_ms(), n.duration_ms.max(1)),
-        None => (0, 1),
-    };
-    // Compute the bar width from the exact label lengths so the duration sits
-    // flush against the right edge (aligned with the volume meter above it).
-    let left = format!("{} ", fmt_ms(pos));
-    let right = format!(" {}", fmt_ms(dur));
-    let reserve = left.chars().count() + right.chars().count();
-    let bar_w = (area.width as usize).saturating_sub(reserve);
-    let filled = ((pos as f32 / dur as f32) * bar_w as f32) as usize;
-
-    let mut spans = vec![Span::styled(left, theme.muted())];
-    spans.extend(gradient_progress(
-        bar_w,
-        filled,
-        &[theme.primary, theme.accent],
-        theme.border_dimmest,
-    ));
-    spans.push(Span::styled(right, theme.muted()));
-    f.render_widget(Paragraph::new(Line::from(spans)), area);
-}
-
-fn render_footer(f: &mut Frame, app: &App, theme: Theme, area: Rect) {
-    let on = |b: bool| if b { theme.success } else { theme.text_muted };
-    let key = |k: &'static str| Span::styled(k, Style::default().fg(theme.primary.into()));
-    let lbl = |t: &'static str| Span::styled(t, theme.muted());
-    let enter_lbl = enter_label(app.cur_items().get(app.browse.selected));
-    let play_lbl = if app.playback.now.as_ref().is_some_and(|n| n.is_playing) {
-        " pause   "
-    } else {
-        " play    "
-    };
-    // Flagged hints go with the library pane, which zen hides — a key that does
-    // nothing must not be advertised.
-    let hints = [
-        (true, key("⇥"), lbl(" section   ")),
-        (false, key("←→"), lbl(" view   ")),
-        (true, key("/"), lbl(" search   ")),
-        (
-            true,
-            key("⏎"),
-            Span::styled(format!(" {enter_lbl}   "), theme.muted()),
-        ),
-        (true, key("⇧⏎"), lbl(" play   ")),
-        (true, key("S"), lbl(" shuffle   ")),
-        (false, key("␣"), Span::styled(play_lbl, theme.muted())),
-        (false, key("n/b"), lbl(" skip   ")),
-        (false, key("⇧←→"), lbl(" seek   ")),
-        (true, key("o"), lbl(" sort   ")),
-        (false, key("+/-"), lbl(" vol   ")),
-        (
-            false,
-            Span::styled("s", Style::default().fg(on(app.transport.shuffle).into())),
-            lbl(" shuffle   "),
-        ),
-        (false, key("a"), lbl(" actions   ")),
-        (
-            false,
-            Span::styled("z", Style::default().fg(on(app.view.zen).into())),
-            lbl(" zen   "),
-        ),
-        (false, key("q"), lbl(" quit")),
-    ];
-    let line = Line::from(
-        hints
-            .into_iter()
-            .filter(|(needs_library, ..)| !needs_library || !app.view.zen)
-            .flat_map(|(_, k, label)| [k, label])
-            .collect::<Vec<_>>(),
-    );
-    f.render_widget(Paragraph::new(line).alignment(Alignment::Center), area);
-}
-
-/// Context actions menu — a centered overlay list.
-fn render_actions_overlay(f: &mut Frame, app: &App, theme: Theme, area: Rect) {
-    let Some(menu) = &app.view.actions else {
-        return;
-    };
-    let w = (area.width * 5 / 10).clamp(28, 52);
-    let h = (menu.items.len() as u16 + 4).clamp(6, area.height.saturating_sub(2));
-    let x = area.x + area.width.saturating_sub(w) / 2;
-    let y = area.y + area.height.saturating_sub(h) / 2;
-    let rect = Rect {
-        x,
-        y,
-        width: w,
-        height: h,
-    };
-
-    f.render_widget(Clear, rect);
-    f.render_widget(Block::default().style(theme.element()), rect);
-    let inner = rect.inner(Margin::new(2, 1));
-    let max = inner.width as usize;
-    let mut lines = vec![
-        Line::from(Span::styled(truncate(&menu.title, max), theme.heading())),
-        Line::raw(""),
-    ];
-    for (i, it) in menu
-        .items
-        .iter()
-        .take(inner.height.saturating_sub(2) as usize)
-        .enumerate()
-    {
-        if i == menu.selected {
-            lines.push(Line::from(vec![
-                Span::styled("› ", Style::default().fg(theme.primary.into())),
-                Span::styled(
-                    truncate(&it.label, max.saturating_sub(2)),
-                    Style::default()
-                        .fg(theme.text.into())
-                        .add_modifier(Modifier::BOLD),
-                ),
-            ]));
-        } else {
-            lines.push(Line::from(Span::styled(
-                format!("  {}", truncate(&it.label, max.saturating_sub(2))),
-                theme.muted(),
-            )));
-        }
-    }
-    f.render_widget(Paragraph::new(lines), inner);
-    force_area(f, rect);
-}
-
-fn render_queue_view(f: &mut Frame, app: &App, theme: Theme, area: Rect) {
-    let inner = area.inner(Margin::new(2, 1));
-    if inner.height == 0 {
-        return;
-    }
-    let max = inner.width as usize;
-    let mut lines: Vec<Line> = Vec::new();
-
-    // Context header — what's playing from.
-    if !app.transport.source_name.is_empty() {
-        lines.push(Line::from(vec![
-            Span::styled("PLAYING FROM  ", theme.muted()),
-            Span::styled(
-                truncate(&app.transport.source_name, max.saturating_sub(14)),
-                Style::default()
-                    .fg(theme.primary.into())
-                    .add_modifier(Modifier::BOLD),
-            ),
-        ]));
-        lines.push(Line::raw(""));
-    }
-
-    // Now playing — the current track, above the up-next list.
-    if let Some(n) = app.playback.now.as_ref() {
-        lines.push(Line::from(Span::styled("NOW PLAYING", theme.heading())));
-        lines.push(Line::from(vec![
-            Span::styled("   ", theme.muted()),
-            Span::styled(
-                truncate(&n.title, max.saturating_sub(3)),
-                Style::default()
-                    .fg(theme.text.into())
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::styled(format!("  {}", n.artist), theme.muted()),
-        ]));
-        lines.push(Line::raw(""));
-    }
-
-    lines.push(Line::from(Span::styled("UP NEXT", theme.heading())));
-    lines.push(Line::raw(""));
-
-    let used = lines.len();
-    if app.transport.queue.is_empty() {
-        lines.push(Line::from(Span::styled("queue is empty", theme.muted())));
-    } else {
-        for (i, q) in app
-            .transport
-            .queue
-            .iter()
-            .take(inner.height.saturating_sub(used as u16) as usize)
-            .enumerate()
-        {
-            lines.push(Line::from(vec![
-                Span::styled(format!("{:>2}  ", i + 1), theme.muted()),
-                Span::styled(
-                    truncate(q, max.saturating_sub(4)),
-                    Style::default().fg(theme.text.into()),
-                ),
-            ]));
-        }
-    }
-    f.render_widget(Paragraph::new(lines), inner);
-}
-
 /// A blocking HTTP client with a timeout so a stalled network can't wedge a
 /// worker thread forever (audit H2).
 fn http_client() -> reqwest::blocking::Client {
@@ -4514,8 +3570,6 @@ fn http_client() -> reqwest::blocking::Client {
 }
 
 // ------------------------------------------------------------------ terminal
-
-const SPINNER: [&str; 8] = ["⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧"];
 
 /// Run `task`, drawing the startup screen until it finishes.
 async fn with_loader<T>(
@@ -4536,41 +3590,6 @@ async fn with_loader<T>(
             }
         }
     }
-}
-
-/// The startup screen: wordmark, spinner, and what we're waiting on.
-fn render_loading(f: &mut Frame, label: &str, frame: usize) {
-    let theme = TOKYONIGHT;
-    let area = f.area();
-    f.render_widget(Block::default().style(theme.panel()), area);
-
-    let top = area.y + area.height.saturating_sub(3) / 2;
-    let row = |dy: u16| Rect {
-        x: area.x,
-        y: top.saturating_add(dy).min(area.bottom().saturating_sub(1)),
-        width: area.width,
-        height: 1,
-    };
-
-    let mark: Vec<Span> = gradient_line("\u{FF2D}\u{FF39}\u{FF38}", &[theme.primary, theme.accent])
-        .into_iter()
-        .map(|mut sp| {
-            sp.style = sp.style.add_modifier(Modifier::BOLD);
-            sp
-        })
-        .collect();
-    f.render_widget(
-        Paragraph::new(Line::from(mark)).alignment(Alignment::Center),
-        row(0),
-    );
-    f.render_widget(
-        Paragraph::new(Line::from(vec![
-            Span::styled(SPINNER[frame % SPINNER.len()], theme.heading()),
-            Span::styled(format!("  {label}…"), theme.muted()),
-        ]))
-        .alignment(Alignment::Center),
-        row(2),
-    );
 }
 
 // ------------------------------------------------------------------ tests
