+++ src/main_tests/nav.rs
@@ -1,3 +1,4 @@
+use crate::ui::*;
 use crate::*;
 use ratatui::Terminal;
